refactor command location
